+++ token-client/src/main/java/com/sap/cloud/security/xsuaa/client/OAuth2ServiceException.java
@@ -8,6 +8,9 @@ package com.sap.cloud.security.xsuaa.client;
 import java.io.IOException;
 import java.io.Serial;
 import java.net.URI;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.List;
 import java.util.Objects;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
@@ -20,6 +23,7 @@ public class OAuth2ServiceException extends IOException {
 	@Serial
 	private static final long serialVersionUID = 1L;
 	private Integer httpStatusCode = 0;
+	private final List<String> headers = new ArrayList<>();
 
 	public OAuth2ServiceException(String message) {
 		super(message);
@@ -43,6 +47,40 @@ public class OAuth2ServiceException extends IOException {
 	 *
 	 * @param message
 	 *            the error message
+	 * @param httpStatusCode
+	 *            the status code of the HTTP service request
+	 * @param headers
+	 * 	          the headers of the HTTP service request
+	 */
+
+	OAuth2ServiceException(String message, Integer httpStatusCode, List<String> headers) {
+		this(message, httpStatusCode);
+		this.headers.addAll(headers);
+	}
+
+	/**
+	 * Returns the HTTP status code of the failed OAuth2 service request or
+	 * {@code 0} e.g. in case the service wasn't called at all.
+	 *
+	 * @return status code or 0
+	 */
+	public Integer getHttpStatusCode() {
+		return httpStatusCode;
+	}
+
+	/**
+	 * Returns the HTTP headers of the failed OAuth2 service request
+	 * @return list of HTTP headers
+	 */
+	public List<String> getHeaders() {
+		return this.headers;
+	}
+
+	/**
+	 * Creates an exception.
+	 *
+	 * @param message
+	 *            the error message
 	 */
 	public static Builder builder(String message) {
 		return new Builder(message);
@@ -53,7 +91,8 @@ public class OAuth2ServiceException extends IOException {
 		private Integer httpStatusCode;
 		private URI serverUri;
 		private String responseBody;
-		private String headers;
+		private final List<String> headers = new ArrayList<>();
+		private String headersString;
 
 		public Builder(String message) {
 			this.message = message;
@@ -82,21 +121,21 @@ public class OAuth2ServiceException extends IOException {
 		}
 
 		public Builder withHeaders(String... headers) {
-			this.headers = "[";
-			for (String header : headers) {
-				this.headers += header;
-			}
-			this.headers += "]";
+			List<String> headerList = Arrays.stream(headers).filter(Objects::nonNull).toList();
+
+			this.headers.addAll(headerList);
+			this.headersString = headerList.stream().collect(Collectors.joining(", ", "[", "]"));
+
 			return this;
 		}
 
 		public OAuth2ServiceException build() {
-			String message = Stream
+			String m = Stream
 					.of(this.message, createUriMessage(), createStatusCodeMessage(), createResponseBodyMessage(),
 							createHeaderMessage())
 					.filter(Objects::nonNull)
 					.collect(Collectors.joining(". "));
-			return new OAuth2ServiceException(message, httpStatusCode);
+			return new OAuth2ServiceException(m, httpStatusCode, headers);
 		}
 
 		private String createResponseBodyMessage() {
@@ -112,7 +151,7 @@ public class OAuth2ServiceException extends IOException {
 		}
 
 		private String createHeaderMessage() {
-			return headers == null ? null : "Headers " + headers;
+			return headersString == null ? null : "Headers " + headersString;
 		}
 	}
 }